refactor(bios): moved smbios from bios
